rework Table e2e tests to remove flakiness
There were two e2e tests that were flaky in
`test/e2e/tests/test_table.py`.

The flakiest test was test_enable_ttl, which often would result in the
following:

```
>       assert ttl["TimeToLiveDescription"]["AttributeName"] == "ForumName"
E       KeyError: 'AttributeName'

tests/test_table.py:158: KeyError
```

This was caused by the dynamodb client `describe_time_to_live` returning
a `ResourceNotFoundException` when the Table was not in ACTIVE status
and even when the Table was in ACTIVE status, there is an amount of time
when the DescribeTimeToLive's TimeToLiveDescription output shape is
either empty or is a struct with no fields :(

I updated the test to first wait until the Table becomes ACTIVE and has
valid output responses before attempting to read the TTL information.

The second flaky test was the test_table_update_tags. The sleep after
calling the `kubectl patch` with the updated tags was too short at 5
seconds since it sometimes takes a little while for the tags to appear
in the Tagris APIs. Increasing this sleep to 10 fixes this flakiness.

Issue aws-controllers-k8s/community#1561

Signed-off-by: Jay Pipes <jaypipes@gmail.com>