release artifacts for `v1.0.1` (#64)
Description of changes:
Regenerates the controller as `v1.0.1` to include the code-generator `v0.24.2` changes. When the `v1.0.0` tag was pushed, there was no release made because of the CD cluster running out of pod resources. Therefore, `v1.0.0` was never released. I pushed out the `v1.0.0` tag again, but in that time the `v0.24.2` code-generator update pushed the version from `v0.1.9` to `v0.1.10` and included those changes.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.