Reuse UpdateGlobalTable for both Update and Delete operations (#92)
It looks like to delete a GlobalTable we need to delete all it's region
replications. Annd with the new code-generator feature, adding support
of assigning the same AWS operation to multiple resources and CRUD
types. We'll have to explicitely set that this operation supports both
update and delete.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.