make `tableReplica` a no-op field (#124)
Previously, we released a new patch version of dynamodb that introduced a new field called tableReplicas. 

It turns out that updating to this change could trigger accidental deletion of table replicas that were previously/manually created or managed with different tools (such as CDK/CF). To smoothen the upgrade process, we propose to revert this change in a patch release, and re-introduce tableReplicas in a minor one, aligning more with semantic versioning standards and signaling to ACK users that a significant feature has been introduced so that they can read the change log and take the necessary actions.

We also propose that we create a standard/process around versioning, ideally new fields needs to be introduced in minor releases, patches should only include bug fixes and non-break behavior changes.

Description of changes:
- removing tableReplica hooks , will be included with a minor release.
- `tableReplica` field remains as a no-op field

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.