fix: change syncTableTags parameter order to avoid panic
In syncTableTags, there was an issue with the order/naming of desired vs
latest resource.

Although the function is called using `syncTableTags(desired, latest)`
the function definition had `syncTableTags(latest, desired)`. This
becomes an issue when we try to retrieve the ARN from latest, during
adoption.
Although in the funciton we named the parameter `latest`, this parameter
was treated as the desired resource when computing the tags delta.

During adoption, the Status/ARN of the desired resource is meant to be
nil, unitl it is later populated. Hence our need to use the ARN from
latest