fix: change syncTableTags parameter order to avoid panic (#136)
Issue [#2459#issuecomment-2886623397](https://github.com/aws-controllers-k8s/community/issues/2459#issuecomment-2886623397)

Description of changes:
In syncTableTags, there was an issue with the order/naming of desired vs
latest resource.

Although the function is called using `syncTableTags(desired, latest)`
the function definition had `syncTableTags(latest, desired)`. This
becomes an issue when we try to retrieve the ARN from latest, during
adoption.
Although in the funciton we named the parameter `latest`, this parameter
was treated as the desired resource when computing the tags delta.

During adoption, the Status/ARN of the desired resource is meant to be 
nil, unitl it is later populated. Hence our need to use the ARN from 
latest.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.