chore: move updates from create to sdkUpdate (#141)
Description of changes:
Currently, we attempt to make certain update calls in sdkCreate, right
after we populate the resource from createOutput. These calls usually
fail as the Table may still be in `Creating` status.

Another issue that kept happening was a throttling error during sdkFind
for the DescribeTTL call. With these changes, we will be checking the
Table status and reconciling if needed, and only retrieving TTL when the
table has `available` status.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.